test: Move tests for PlotDataPoint into new test module
PlotDataPoint is a general class used troughout all "quant" modules,
it doesn't need to be tested with the output of different software
tools or for different "quant" modules.
